Phase K + Phase R wave 4: sweep PacketStructures.h long  int32_t
Completes the long  int32_t migration for the remaining ~152 wire-struct
field declarations in common/include/net7/PacketStructures.h. Prior
waves (commit e74f07c then 2026-05-24 GlobalAvatarList family then
ColorInfo/ShipData/GlobalCreateCharacter) migrated the structs actively
consumed by Linux Phase J/K handlers; this sweep covers everything
else so future opcode handlers landing on Linux don't trip the same
Win32-4B-vs-Linux-8B wire-format drift on first encounter.

Mechanical: every `long Foo[;,[]` and `unsigned long Foo[;,[]` line
inside ATTRIB_PACKED structs becomes `int32_t Foo...` / `uint32_t Foo...`.
The header has no function signatures and no local variables — only
struct fields and #defines — so a per-line `^\s*(unsigned\s+)?long\s+\w`
substitution converts exactly the wire-field declarations without
touching the existing migration comments.

Structs migrated this commit (covering opcodes the proxy/server stub
out today but will need when their handlers land): WarpPacket / InvMove
/ InvSort / ItemState / ChangeBaseAsset / Create / ServerParameters
(DebrisMode + SectorNum) / Login (TimeSent + TimeReceived) / Navigation
/ CreateAttachment / DecalItem / Decal / NameDecal / ColorizationItem /
Colorization / AvatarDescription / Subparts / ConstantPositionalUpdate
/ FormationPositionalUpdate / RequestTarget / SetInterface / SetTarget
/ ActionPacket / ActionPacket2 / ClientSetTime / VerbRequest /
CameraControl / LogoffRequest / TriggerEmote / NotifyEmote /
OptionPacket / SelectTalkTree / ChatStream / ClientChat /
ClientChatRequest / ClientChatList / ClientChatError / ClientChatEvent
/ ClientSkillsRequest / StarbaseAvatarChange{,_S2C} / StarbaseRoomChange
/ StarbaseRequest / Relationship / ObjectEffect / ObjectToObjectEffect
/ InitRenderState / ActivateRenderState / SimplePositionalUpdate /
PlanetPositionalUpdate / ComponentPositionalUpdate /
AdvancedPositionalUpdate / EquipUse / AbilityUse / StarbaseSet /
ShipInfo / CTARequest / MovePacket / SkillAction / SkillUse /
MissionDismissal / MVASHandoff / ManufactureData /
ManufactureTechLevelFilter / FindMember (+ nested fm_item) /
RecustomizeAvatarStart / RecustomizeShipStart / RecustomizeShipDone /
RecustomizeAvatarDone.

Knock-on fix: `Player::SetupWarpNavs(short, long *)` decl/defn was
referenced from PlayerConnection.cpp's HandleWarp at two call sites
that now pass `int32_t (&)[20]` (WarpPacket.TargetID). Updated the
signature in PlayerClass.h + PlayerClass.cpp to take `int32_t *target_id`
so the function matches the migrated wire-struct type. The function
body itself is fine — it just assigns into `m_WarpNavs[i]` which is
a long member of Player (unrelated to wire format).

`int32_t` and `unsigned long` -> `uint32_t` swaps are value-preserving
at every assignment site (ntohl() returns uint32_t, sprintf %ld
%d, etc.) and ATTRIB_PACKED still guards the no-padding invariant.

Build verification: server + proxy + login-server all compile clean
against the migrated header. 38/38 cli-integration tests green from a
fresh `docker compose down -v` rebuild.

+++ server/src/PlayerClass.h
@@ -388,7 +388,7 @@ public:
     Object     *GetNearestNav()                 { return (m_NearestNav); }
 
     //warp methods
-    void        SetupWarpNavs(short navs, long *target_id);
+    void        SetupWarpNavs(short navs, int32_t *target_id);
     void        PrepareForWarp();
     void        StartWarp();
     void        TerminateWarp(bool player_forced = false);
